@@ -7,8 +7,8 @@ set(NEML_version 1.1.0)
 ### Setup modules ###
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/cmake/Modules/")
 
-### Need c++ 11 ###
-set(CMAKE_CXX_STANDARD 11)
+### Need c++ 17 ###
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 ### Default to shared libs ###
@@ -30,10 +30,6 @@ elseif (${SOLVER} MATCHES "nox")
 else()
       message("Unknown solver")
 endif()
-
-### Need Boost (only variant.hpp really) ###
-FIND_PACKAGE(Boost REQUIRED)
-INCLUDE_DIRECTORIES(SYSTEM ${Boost_INCLUDE_DIRS})
 
 ### Configure standard-ish libraries ###
 set(BLAS_VENDOR Generic)
