@@ -105,7 +105,7 @@ std::unique_ptr<NEMLObject> get_object_unique(const rapidxml::xml_node<> * node)
   // Special case: could be a ConstantInterpolate
   std::string type = get_type_of_node(node);
   if (type == "none") {
-    return neml::make_unique<ConstantInterpolate>(get_double(node));
+    return std::make_unique<ConstantInterpolate>(get_double(node));
   }
   else {
     ParameterSet params = get_parameters(node);
@@ -208,7 +208,7 @@ std::vector<std::shared_ptr<NEMLObject>> get_vector_object(
       (node->first_node()->type() == rapidxml::node_data)) {
     std::vector<double> data = get_vector_double(node);
     for (auto v : data) {
-      joined.push_back(neml::make_unique<ConstantInterpolate>(v));
+      joined.push_back(std::make_unique<ConstantInterpolate>(v));
     }
     return joined;
   }
