@@ -17,7 +17,7 @@ std::string IsoJ2::type()
 
 std::unique_ptr<NEMLObject> IsoJ2::initialize(ParameterSet & params)
 {
-  return neml::make_unique<IsoJ2>(); 
+  return std::make_unique<IsoJ2>(); 
 }
 
 ParameterSet IsoJ2::parameters()
@@ -45,7 +45,7 @@ std::string IsoKinJ2::type()
 
 std::unique_ptr<NEMLObject> IsoKinJ2::initialize(ParameterSet & params)
 {
-  return neml::make_unique<IsoKinJ2>(); 
+  return std::make_unique<IsoKinJ2>(); 
 }
 
 ParameterSet IsoKinJ2::parameters()
@@ -238,7 +238,7 @@ std::string IsoKinJ2I1::type()
 
 std::unique_ptr<NEMLObject> IsoKinJ2I1::initialize(ParameterSet & params)
 {
-  return neml::make_unique<IsoKinJ2I1>(
+  return std::make_unique<IsoKinJ2I1>(
       params.get_object_parameter<Interpolate>("h"),
       params.get_object_parameter<Interpolate>("l")
       ); 
@@ -470,7 +470,7 @@ std::string IsoJ2I1::type()
 
 std::unique_ptr<NEMLObject> IsoJ2I1::initialize(ParameterSet & params)
 {
-  return neml::make_unique<IsoJ2I1>(
+  return std::make_unique<IsoJ2I1>(
       params.get_object_parameter<Interpolate>("h"),
       params.get_object_parameter<Interpolate>("l")
       ); 
