@@ -132,7 +132,7 @@ ParameterSet GPowerLaw::parameters()
 
 std::unique_ptr<NEMLObject> GPowerLaw::initialize(ParameterSet & params)
 {
-  return neml::make_unique<GPowerLaw>(
+  return std::make_unique<GPowerLaw>(
       params.get_object_parameter<Interpolate>("n"),
       params.get_object_parameter<Interpolate>("eta")
       ); 
@@ -185,7 +185,7 @@ ParameterSet PerzynaFlowRule::parameters()
 
 std::unique_ptr<NEMLObject> PerzynaFlowRule::initialize(ParameterSet & params)
 {
-  return neml::make_unique<PerzynaFlowRule>(
+  return std::make_unique<PerzynaFlowRule>(
       params.get_object_parameter<YieldSurface>("surface"),
       params.get_object_parameter<HardeningRule>("hardening"),
       params.get_object_parameter<GFlow>("g")
@@ -404,7 +404,7 @@ ParameterSet ConstantFluidity::parameters()
 
 std::unique_ptr<NEMLObject> ConstantFluidity::initialize(ParameterSet & params)
 {
-  return neml::make_unique<ConstantFluidity>(
+  return std::make_unique<ConstantFluidity>(
       params.get_object_parameter<Interpolate>("eta")
       ); 
 }
@@ -446,7 +446,7 @@ ParameterSet SaturatingFluidity::parameters()
 
 std::unique_ptr<NEMLObject> SaturatingFluidity::initialize(ParameterSet & params)
 {
-  return neml::make_unique<SaturatingFluidity>(
+  return std::make_unique<SaturatingFluidity>(
       params.get_object_parameter<Interpolate>("K0"),
       params.get_object_parameter<Interpolate>("A"),
       params.get_object_parameter<Interpolate>("b")
@@ -500,7 +500,7 @@ ParameterSet ChabocheFlowRule::parameters()
 
 std::unique_ptr<NEMLObject> ChabocheFlowRule::initialize(ParameterSet & params)
 {
-  return neml::make_unique<ChabocheFlowRule>(
+  return std::make_unique<ChabocheFlowRule>(
       params.get_object_parameter<YieldSurface>("surface"),
       params.get_object_parameter<NonAssociativeHardening>("hardening"),
       params.get_object_parameter<FluidityModel>("fluidity"),
@@ -731,7 +731,7 @@ ParameterSet YaguchiGr91FlowRule::parameters()
 
 std::unique_ptr<NEMLObject> YaguchiGr91FlowRule::initialize(ParameterSet & params)
 {
-  return neml::make_unique<YaguchiGr91FlowRule>(); 
+  return std::make_unique<YaguchiGr91FlowRule>(); 
 }
 
 
